fix: refactor code to use foreach_collect
